refactor(dashboard): adopt reinhardt::pages::ClientLauncher in WASM entry
Replace the manual WASM bootstrap in `dashboard/src/client.rs` with
`reinhardt::pages::ClientLauncher`, which already handles router init,
history listener wiring, DOM mount on `#app`, and the reactive re-render
Effect. Drop the duplicated `thread_local! ROUTER` plus
`init_global_router` / `with_router` from `dashboard/src/client/router.rs`
and route runtime access through `reinhardt::pages::with_router`. Extract
the SPA `<a>` click delegation into a dedicated
`dashboard/src/client/link_interception.rs` module so each module owns
exactly one concern, in line with the dashboard's MF-1 (aggregator-only)
and one-component-per-file conventions.

`ClientLauncher` does not install link interception in
reinhardt-web v0.1.0-rc.22, so the dashboard continues to own that
small piece. The toast container insert and `ws::connect_notifications`
are kept as a path-driven side-effect Effect installed after `launch()`
so authenticated routes still bring up notifications without changing
behaviour.

Fixes #442
